Title: 달에서 무게 구하기, Time: 96ms, Memory: 9MB, Status: Passed - Codetree

+++ 260218/달에서 무게 구하기/weight-on-the-moon.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
         // Please write your code here.
         int weight = 13;
-        double ratio = 0.165;
+        float ratio = 0.165f;
 
         System.out.printf("%d * %.6f = %.6f", weight, ratio, weight*ratio);
     }
